requirements.txt
discord.py==2.5.0
parfive
requests